feat: implement smooth player movement
